fix update profile

+++ frontend/src/api/user.js
@@ -1,13 +1,6 @@
 import API from './axios';
 
-// FR-03: Update profil
-export const updateProfil = (userId, data) =>
-  API.put(`/user/update/${userId}`, data);
-
-// FR-04: Hapus akun
-export const deleteAkun = (userId) =>
-  API.delete(`/user/delete/${userId}`);
-
-// FR-06: Logout
-export const logout = () =>
-  API.post('/user/logout');
+export const getProfil = (id) => API.get(`/user/profile/${id}`);
+export const updateProfil = (userId, data) => API.put(`/user/update/${userId}`, data);
+export const deleteAkun = (userId) => API.delete(`/user/delete/${userId}`);
+export const logout = () => API.post('/user/logout');
